+++ src/main/java/com/lb/lightboard/controller/user/UserController.java
@@ -1,11 +1,8 @@
 package com.lb.lightboard.controller.user;
 
 import com.lb.lightboard.bo.UserBO;
-import com.lb.lightboard.controller.CrudController;
-import com.lb.lightboard.model.entity.User;
 import com.lb.lightboard.model.network.Header;
-import com.lb.lightboard.model.network.request.UserApiRequest;
-import com.lb.lightboard.model.network.response.UserApiResponse;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,15 +10,18 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping(value = "/v1/api/user", produces = MediaType.APPLICATION_JSON_VALUE)
-public class UserController extends CrudController<UserApiRequest, UserApiResponse, User> {
+@RequestMapping(value = "/v1/api/admin/user", produces = MediaType.APPLICATION_JSON_VALUE)
+public class UserController {
+	@Autowired
+	UserBO userBO;
+	
 	@GetMapping(params = "userId")
 	public Header<Boolean> isDuplicateUserId(@RequestParam(value = "userId") String userId) {
-		return ((UserBO)baseBO).isExistUserId(userId);
+		return userBO.isExistUserId(userId);
 	}
 	
 	@GetMapping(value = "/admin/check")
 	public Header<Boolean> isExistAdminUser() {
-		return ((UserBO)baseBO).isExistAdminUser();
+		return userBO.isExistAdminUser();
 	}
 }
